# src/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Profile": {
				"description": "A profile is a specific customisation for each endpoint.\nIt parses data from the API differently, add additional information, or enable non-default methods.",
				"properties": {
					"locale": {
						"type": "string"
					},
					"timezone": {
						"type": "string"
					},
					"endpoint": {
						"type": "string"
					},
					"products": {
						"items": {
							"$ref": "#/components/schemas/ProductType"
						},
						"type": "array"
					},
					"trip": {
						"type": "boolean"
					},
					"radar": {
						"type": "boolean"
					},
					"refreshJourney": {
						"type": "boolean"
					},
					"journeysFromTrip": {
						"type": "boolean"
					},
					"reachableFrom": {
						"type": "boolean"
					},
					"journeysWalkingSpeed": {
						"type": "boolean"
					},
					"tripsByName": {
						"type": "boolean"
					},
					"remarks": {
						"type": "boolean"
					},
					"remarksGetPolyline": {
						"type": "boolean"
					},
					"lines": {
						"type": "boolean"
					}
				},
				"required": [
					"locale",
					"timezone",
					"endpoint",
					"products"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProfileId": {
				"type": "string",
				"enum": [
					"db",
					"bvg",
					"oebb"
				]
			},
...
			"Location": {
				"description": "A location object is used by other items to indicate their locations.",
				"properties": {
					"type": {
						"type": "string",
						"enum": [
							"location"
						],
						"nullable": false
					},
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"poi": {
						"type": "boolean"
					},
					"address": {
						"type": "string"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"altitude": {
						"type": "number",
						"format": "double"
					},
					"distance": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Journeys": {
				"properties": {
					"realtimeDataUpdatedAt": {
						"type": "number",
						"format": "double"
					},
					"earlierRef": {
						"type": "string"
					},
					"laterRef": {
						"type": "string"
					},
					"journeys": {
						"items": {
							"$ref": "#/components/schemas/Journey"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"TripWithRealtimeData": {
				"properties": {
					"realtimeDataUpdatedAt": {
						"type": "number",
						"format": "double"
					},
					"trip": {
						"$ref": "#/components/schemas/Trip"
					}
				},
				"required": [
					"trip"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Departures": {
				"properties": {
					"realtimeDataUpdatedAt": {
						"type": "number",
						"format": "double"
					},
					"departures": {
						"items": {
							"$ref": "#/components/schemas/Alternative"
						},
						"type": "array"
					}
				},
				"required": [
					"departures"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "hafas-client-rest-api",
		"version": "0.0.1",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/hafas/{profileId}/profile": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Profile"
								}
							}
						}
					}
				},
				"description": "Retrieves profile",
				"security": [],
				"parameters": [
					{
						"description": "endpoint profile id",
						"in": "path",
						"name": "profileId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ProfileId"
						}
					}
				]
			}
		},
		"/hafas/{profileId}/journeys": {
			"get": {
				"operationId": "Getjourneys",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Journeys"
								}
							}
						}
					}
				},
				"description": "Retrieves journeys",
				"security": [],
				"parameters": [
					{
						"description": "endpoint profile id",
						"in": "path",
						"name": "profileId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ProfileId"
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "departure date, undefined corresponds to Date.Now",
						"in": "query",
						"name": "departure",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "how many search results?",
						"in": "query",
						"name": "results",
						"required": false,
						"schema": {
							"default": 3,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "let journeys pass this station",
						"in": "query",
						"name": "via",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "return stations on the way?",
						"in": "query",
						"name": "stopovers",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "Maximum nr of transfers. Default: Let HAFAS decide.",
						"in": "query",
						"name": "transfers",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "minimum time for a single transfer in minutes",
						"in": "query",
						"name": "transferTime",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "'none', 'partial' or 'complete'",
						"in": "query",
						"name": "accessibility",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "only bike-friendly journeys",
						"in": "query",
						"name": "bike",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "return tickets? only available with some profiles",
						"in": "query",
						"name": "tickets",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "return a shape for each leg?",
						"in": "query",
						"name": "polylines",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose sub-stops of stations?",
						"in": "query",
						"name": "subStops",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose entrances of stops/stations?",
						"in": "query",
						"name": "entrances",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose hints & warnings?",
						"in": "query",
						"name": "remarks",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "'slow', 'normal', 'fast'",
						"in": "query",
						"name": "walkingSpeed",
						"required": false,
						"schema": {
							"default": "slow",
							"type": "string"
						}
					},
					{
						"description": "start with walking",
						"in": "query",
						"name": "startWithWalking",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "language to get results in",
						"in": "query",
						"name": "language",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "parse which days each journey is valid on",
						"in": "query",
						"name": "scheduledDays",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			}
		},
		"/hafas/{profileId}/trip": {
			"get": {
				"operationId": "Gettrip",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TripWithRealtimeData"
								}
							}
						}
					}
				},
				"description": "Refetch information about a trip",
				"security": [],
				"parameters": [
					{
						"description": "endpoint profile id",
						"in": "path",
						"name": "profileId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ProfileId"
						}
					},
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "return stations on the way?",
						"in": "query",
						"name": "stopovers",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "return a shape for the trip?",
						"in": "query",
						"name": "polyline",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose sub-stops of stations?",
						"in": "query",
						"name": "subStops",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose entrances of stops/stations?",
						"in": "query",
						"name": "entrances",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose hints & warnings?",
						"in": "query",
						"name": "remarks",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "parse which days each journey is valid on",
						"in": "query",
						"name": "scheduledDays",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "Language of the results",
						"in": "query",
						"name": "language",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				]
			}
		},
		"/hafas/{profileId}/departures": {
			"get": {
				"operationId": "Getdepartures",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Departures"
								}
							}
						}
					}
				},
				"description": "Retrieves departures",
				"security": [],
				"parameters": [
					{
						"description": "endpoint profile id",
						"in": "path",
						"name": "profileId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ProfileId"
						}
					},
					{
						"in": "query",
						"name": "station",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "departure date, undefined corresponds to Date.Now",
						"in": "query",
						"name": "when",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "only show departures heading to this station",
						"in": "query",
						"name": "direction",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "filter by line ID",
						"in": "query",
						"name": "line",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "show departures for the next n minutes",
						"in": "query",
						"name": "duration",
						"required": false,
						"schema": {
							"default": 120,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "max. number of results; `null` means \"whatever HAFAS wants\"",
						"in": "query",
						"name": "results",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "parse & expose sub-stops of stations?",
						"in": "query",
						"name": "subStops",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose entrances of stops/stations?",
						"in": "query",
						"name": "entrances",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose lines at the stop/station?",
						"in": "query",
						"name": "linesOfStops",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose hints & warnings?",
						"in": "query",
						"name": "remarks",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "products",
						"in": "query",
						"name": "products",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "language",
						"in": "query",
						"name": "language",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				]
			}
		},
		"/hafas/{profileId}/locations": {
			"get": {
				"operationId": "Getlocations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"anyOf": [
											{
												"$ref": "#/components/schemas/Station"
											},
											{
												"$ref": "#/components/schemas/Stop"
											},
											{
												"$ref": "#/components/schemas/Location"
											}
										]
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves locations or stops",
				"security": [],
				"parameters": [
					{
						"description": "endpoint profile id",
						"in": "path",
						"name": "profileId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ProfileId"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "find only exact matches?",
						"in": "query",
						"name": "fuzzy",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "how many search results?",
						"in": "query",
						"name": "results",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "return stops/stations?",
						"in": "query",
						"name": "stops",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "return addresses",
						"in": "query",
						"name": "addresses",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "points of interest",
						"in": "query",
						"name": "poi",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose sub-stops of stations?",
						"in": "query",
						"name": "subStops",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose entrances of stops/stations?",
						"in": "query",
						"name": "entrances",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose lines at each stop/station?",
						"in": "query",
						"name": "linesOfStops",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "Language of the results",
						"in": "query",
						"name": "language",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				]
			}
		},
		"/hafas/{profileId}/stop": {
			"get": {
				"operationId": "Getstop",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Station"
										},
										{
											"$ref": "#/components/schemas/Stop"
										},
										{
											"$ref": "#/components/schemas/Location"
										}
									]
								}
							}
						}
					}
				},
				"description": "Retrieves information about a stop",
				"security": [],
				"parameters": [
					{
						"description": "endpoint profile id",
						"in": "path",
						"name": "profileId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ProfileId"
						}
					},
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "parse & expose lines at the stop/station?",
						"in": "query",
						"name": "linesOfStops",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose sub-stops of stations?",
						"in": "query",
						"name": "subStops",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose entrances of stops/stations?",
						"in": "query",
						"name": "entrances",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose hints & warnings?",
						"in": "query",
						"name": "remarks",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "Language of the results",
						"in": "query",
						"name": "language",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				]
			}
		},
		"/hafas/{profileId}/lines": {
			"get": {
				"operationId": "Getlines",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LinesWithRealtimeData"
								}
							}
						}
					}
				},
				"description": "Fetches all lines known to the HAFAS endpoint",
				"security": [],
				"parameters": [
					{
						"description": "endpoint profile id",
						"in": "path",
						"name": "profileId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ProfileId"
						}
					},
					{
						"in": "query",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Language of the results",
						"in": "query",
						"name": "language",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				]
			}
		},
		"/hafas/{profileId}/nearby": {
			"get": {
				"operationId": "Getnearby",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"anyOf": [
											{
												"$ref": "#/components/schemas/Station"
											},
											{
												"$ref": "#/components/schemas/Stop"
											},
											{
												"$ref": "#/components/schemas/Location"
											}
										]
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves nearby stops from location",
				"security": [],
				"parameters": [
					{
						"description": "endpoint profile id",
						"in": "path",
						"name": "profileId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ProfileId"
						}
					},
					{
						"in": "query",
						"name": "longitude",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "latitude",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "maximum number of results",
						"in": "query",
						"name": "results",
						"required": false,
						"schema": {
							"default": 8,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "maximum walking distance in meters",
						"in": "query",
						"name": "distance",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "return points of interest?",
						"in": "query",
						"name": "poi",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "return stops/stations?",
						"in": "query",
						"name": "stops",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "products",
						"in": "query",
						"name": "products",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "parse & expose sub-stops of stations?",
						"in": "query",
						"name": "subStops",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose entrances of stops/stations?",
						"in": "query",
						"name": "entrances",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"description": "parse & expose lines at each stop/station?",
						"in": "query",
						"name": "linesOfStops",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "language",
						"in": "query",
						"name": "language",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}